+++ api/RunAnalyzer.ps1
@@ -1,1 +1,0 @@
-﻿.\Analyze-PSImmichAPI.ps1 -ApiSpecFile 'api.2.3.1.json' -SourcePath '..\source\Public' -ExclusionConfigFile 'exclusions.json' -ExcludeSkipped