+++ api/RunComparison.ps1
@@ -1,3 +1,0 @@
-﻿.\Compare-OpenApiSpecs.ps1 `
-    -OldSpecPath '.\api.2.2.0.json' `
-    -NewSpecPath '.\api.2.3.1.json'